+++ CHANGELOG.md
@@ -1,11 +1,27 @@
 
 <a name="HEAD"></a>
-## HEAD (2026-09-02)
+## [HEAD](https://github.com/xoctopus/agents/compare/v0.0.1...HEAD) (2026-09-05)
+
+### Chore
+
+* upgrade dependencies
+
+
+<a name="v0.0.1"></a>
+## v0.0.1 (2026-09-05)
+
+### Chore
+
+* prettify output
 
 ### Ci
 
 * update scripts
 * add update bundle makefile entry
+
+### Docs
+
+* **skills:** fix base SKILL.md frontmatter
 
 ### Feat